backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.0
PyJWT>=2.8.0
SQLAlchemy>=2.0.0
